+++ lib/flutter_formx.dart
@@ -1,4 +1,4 @@
-export 'src/form/form_item.dart';
+export 'src/form/form_x_field.dart';
 export 'src/form/formx.dart';
 export 'src/validator/required_field_validator.dart';
 export 'src/validator/validator.dart';