+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
+- [Breaking change] Rename FormItem to FormXField 
+- [Breaking change] Expose the library file only
 - Improve documentation
 - Fix example project setup
-- Expose the library file only
-
 ## v1.0.0
 - Initial release